Make device factory usage more inline with layer factory.
