+++ src/caffe/solver.cpp
@@ -367,7 +367,7 @@ void SGDSolver<Dtype>::ComputeUpdateValue() {
   }
   Dtype momentum = this->param_.momentum();
   Dtype weight_decay = this->param_.weight_decay();
-  Device<Dtype>* device = DeviceFactory<Dtype>::GetDevice();
+  Device<Dtype>* device = GetDevice<Dtype>();
   for (int param_id = 0; param_id < net_params.size(); ++param_id) {
     // Compute the value to history, and then copy them to the blob's diff.
     Dtype local_rate = rate * net_params_lr[param_id];
